+++ common/src/main/kotlin/com/github/cao/awa/conium/event/trigger/ConiumEventTrigger.kt
@@ -4,5 +4,6 @@ import com.github.cao.awa.conium.event.metadata.ConiumEventMetadata
 
 data class ConiumEventTrigger<I: Any, M: ConiumEventMetadata<I>>(
     val callback: (M) -> Unit,
-    val targetIdentity: (I?) -> Boolean
+    val targetIdentity: (I?) -> Boolean,
+    val alwaysCallback: Boolean = false
 )