+++ CHANGE_LOG.md
@@ -1,7 +1,9 @@
-> 1.0.0-alpha11
+> 1.0.0-alpha12
 
 # Conium event
 1. Add classical event handle way (subscriber and metadata).
+2. Add DSL style event handler
+3. Add async event handler supports
 
 # Item templates
 1. 'used_on_block_convert_to'
@@ -9,3 +11,6 @@
 
 # Block register
 1. Early register the block before world-gen state, let blocks able to participate the world generates
+
+# Depends
+1. Remove fluxia mod, build-in the structuring translator